5 环境污染的预测/城市垃圾污染监测系统 - 冲第一小队/requirements.txt
boto3==1.14.60
boto3-stubs==1.14.60.0
botocore==1.17.60
docutils==0.15.2
jmespath==0.10.0
numpy==1.19.2
Pillow==7.2.0
protobuf==3.11.4
python-dateutil==2.8.1
s3transfer==0.3.3
six==1.14.0
urllib3==1.25.10
